refactor: 이름과 개행 수정 및 상수 수치 명확화
