+++ src/main/java/lotto/application/LottoTicketsDto.java
@@ -10,14 +10,14 @@ public class LottoTicketsDto {
     private final List<LottoDto> lottoTickets;
 
     public LottoTicketsDto(LottoTickets lottoTickets) {
-        this.lottoTickets = getLottoTickets(lottoTickets);
+        this.lottoTickets = convertToLottoDtoFromTickets(lottoTickets);
     }
 
     public LottoTicketsDto(List<List<Integer>> lottoTickets) {
-        this.lottoTickets = createLottoTicketsDto(lottoTickets);
+        this.lottoTickets = convertToLottoDto(lottoTickets);
     }
 
-    public List<LottoDto> createLottoTicketsDto(List<List<Integer>> lottoTickets) {
+    public List<LottoDto> convertToLottoDto(List<List<Integer>> lottoTickets) {
         List<LottoDto> lottoTicketDto = new ArrayList<>();
         for (List<Integer> lottoTicket : lottoTickets) {
             lottoTicketDto.add(new LottoDto(lottoTicket));
@@ -26,9 +26,8 @@ public class LottoTicketsDto {
         return lottoTicketDto;
     }
 
-    private List<LottoDto> getLottoTickets(LottoTickets lottoTickets) {
+    private List<LottoDto> convertToLottoDtoFromTickets(LottoTickets lottoTickets) {
         List<LottoDto> lottoTicketDto = new ArrayList<>();
-
         for (Lotto lottoTicket : lottoTickets.getLottoTickets()) {
             lottoTicketDto.add(new LottoDto(lottoTicket.getNumbers()));
         }
