+++ src/main/java/lotto/domain/LottoConfig.java
@@ -3,8 +3,7 @@ package lotto.domain;
 public enum LottoConfig {
     LOTTO_START(1),
     LOTTO_END(45),
-    LOTTO_MAX_NUMBER(6),
-    LOTTO_BONUS_NUMBER(7);
+    LOTTO_MAX_NUMBER(6);
 
     private final int unit;
